set entities , enums , dtos , set ready data for user when we lunch the app (attention db set to drop and recreate in each start of the app)

+++ gesti-bank-backend/src/main/java/net/axel/gestibankbackend/domain/entities/Quote.java
@@ -29,7 +29,7 @@ public class Quote {
 
     private Instant createdAt;
 
-    private Integer totalAmount;
+    private Double totalAmount;
 
     @Enumerated(EnumType.STRING)
     private QuoteStatus status;
